bp-requirements.txt
weakref
types
copy
reprlib
_weakrefset
builtins
operator
browser.ajax
re
itertools
json
codecs
io
sys
sre_parse
time
datetime
sre_constants
genericpath
stat
_collections
random
copyreg
sre_compile
traceback
os
abc
pprint
collections.abc
_weakref
_sre
__future__
_codecs
_ajax
heapq
keyword
_abcoll
_dummy_thread
math
collections
uuid
posix
posixpath
errno
_thread
